Refactoring ColorModel to the correct place in source tree and subsequent calls.
