+++ src/main/java/net/sf/rails/ui/swing/core/GridFieldDynamic1D.java
@@ -1,8 +1,7 @@
 package net.sf.rails.ui.swing.core;
 
 import com.google.common.base.Preconditions;
-
-import net.sf.rails.game.state.ColorModel;
+import net.sf.rails.game.model.ColorModel;
 import net.sf.rails.game.state.Item;
 import net.sf.rails.game.state.Observable;
